+++ DiscreteSegment.cpp
@@ -2,14 +2,14 @@
 
 namespace Lights
 {
-	DiscreteSegment::DiscreteSegment(LedStrip *target, vector<uint> pixels) : MappedSegment(target)
+	DiscreteSegment::DiscreteSegment( LedStrip* target, vector<uint16_t> pixels ) : MappedSegment( target )
 	{
 		numLeds = pixels.size();
 
 		// Initialise the map
-		pixelMap = new uint[numLeds];
+		pixelMap = new uint16_t[ numLeds ];
 
-		uint mapIndex = 0;
+		uint16_t mapIndex = 0;
 		for (auto &pixel : pixels)
 		{
 			pixelMap[mapIndex++] = pixel;
